+++ src/lib/Libattr/attr_fn_arst.c
@@ -438,8 +438,8 @@ int set_arst(
   {
   int	 i;
   int	 j;
-  unsigned long nsize;
-  unsigned long need;
+  int 	 nsize;
+  int 	 need;
   long	 offset;
   char	*pc;
   long	 used;
@@ -469,8 +469,8 @@ int set_arst(
       return(PBSE_INTERNAL);
       }
 
-    need = sizeof(struct array_strings) + (j - 1) * sizeof (char *);
-    pas=(struct array_strings *)malloc(need);
+    need = (int)sizeof(struct array_strings) + (j - 1) * sizeof (char *);
+    pas=(struct array_strings *)malloc((size_t)need);
 
     if (!pas)
       {
@@ -522,7 +522,7 @@ int set_arst(
 
         nsize += nsize / 2;			/* alloc extra space */
 
-        if (!(pas->as_buf = malloc(nsize))) 
+        if (!(pas->as_buf = malloc((size_t)nsize))) 
           {
           pas->as_bufsize = 0;
 
@@ -552,9 +552,9 @@ int set_arst(
         need = pas->as_bufsize + 2 * nsize;  /* alloc new buf */
 
         if (pas->as_buf)
-          pc = realloc(pas->as_buf,need);
+          pc = realloc(pas->as_buf,(size_t)need);
         else
-          pc = malloc(need);
+          pc = malloc((size_t)need);
 
         if (pc == NULL)
           {
@@ -578,9 +578,9 @@ int set_arst(
 
         j = 3 * j / 2;		/* allocate extra     */
 
-        need = sizeof(struct array_strings) + (j - 1) * sizeof(char *);
+        need = (int)sizeof(struct array_strings) + (j - 1) * sizeof(char *);
        
-        newpas = (struct array_strings *)realloc((char *)pas,need);
+        newpas = (struct array_strings *)realloc((char *)pas,(size_t)need);
 
         if (newpas == NULL)
           {
@@ -624,7 +624,7 @@ int set_arst(
 
             need = pas->as_next - pc;
 
-            memcpy(pas->as_string[i],pc,(int)need);
+            memcpy(pas->as_string[i],pc,(size_t)need);
 
             pas->as_next -= nsize;
